requirements.txt
iconsdk~=1.3.2
libeospy~=2.0.1
